requirements.txt
asyncio==3.4.3
aiohttp==3.7.4.post0
aiofile==3.7.1
PyYAML==5.4.1
